+++ kite/src/main/kotlin/com/tang/kite/config/table/DynamicTableProcessor.kt
@@ -1,0 +1,18 @@
+package com.tang.kite.config.table
+
+/**
+ * Dynamic table processor interface
+ *
+ * @author Tang
+ */
+interface DynamicTableProcessor {
+
+    /**
+     * Process table name
+     *
+     * @param tableName table name
+     * @return processed table name
+     */
+    fun process(tableName: String): String
+
+}