+++ kite/src/main/kotlin/com/tang/kite/config/table/TableConfig.kt
@@ -1,0 +1,16 @@
+package com.tang.kite.config.table
+
+/**
+ * Table properties for table configuration.
+ *
+ * @author Tang
+ */
+object TableConfig {
+
+    /**
+     * Dynamic table name processor.
+     */
+    @JvmStatic
+    var dynamicTableName: DynamicTableProcessor? = null
+
+}